Adding Register Route Funtionality
