+++ package.json
@@ -8,6 +8,7 @@
     "email-validator": "^2.0.4",
     "express": "^4.18.2",
     "express-mongo-sanitize": "^2.2.0",
+    "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.1.0",
     "password-validator": "^5.3.0"
   },
